+++ frontend/package.json
@@ -18,6 +18,7 @@
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
     "firebase": "^12.7.0",
+    "framer-motion": "^12.29.0",
     "lucide-react": "^0.562.0",
     "next": "15.0.3",
     "react": "18.3.1",
