+++ frontend/src/app/HomeClient.tsx
@@ -1,3 +1,4 @@
+// src/app/
 "use client";
 
 import { useRouter, useSearchParams } from "next/navigation";
