+++ frontend/src/components/layout/Footer.tsx
@@ -43,6 +43,7 @@ export default function Footer() {
       <br />
       <br />
       <br />
+            <br />
     </footer>
   );
 }
