+++ frontend/src/components/layout/HamburgerMenu.tsx
@@ -11,11 +11,8 @@ import {
   Users,
   MapPin,
   CheckCircle,
-  LocateFixed,
-  Calculator,
   LogOut,
   ChevronRight,
-  BookA,
 } from 'lucide-react'
 
 interface HamburgerMenuProps {
@@ -30,11 +27,7 @@ const MENUS = [
   { label: '추천 장소 선택', href: '/meetings/new/step3-result', icon: MapPin },
   { label: '모임 확정', href: '/meetings/complete', icon: CheckCircle },
 ]
-const EXTRA_MENUS = [
-  { label: '지도 서비스', href: '/meetings/option-location', icon: LocateFixed },
-  { label: '회비 계산기', href: '/meetings/option-fee', icon: Calculator },
-]
-const ABOUT_MENUS = [{ label: '만날각 소개', href: '/about', icon: BookA }]
+
 
 const isValidUuid = (value: string | null): value is string => {
   if (!value) return false
@@ -187,44 +180,7 @@ const withMeetingUuid = (href: string) => {
         </nav>
 
         <div className="my-4 border-t" />
-
-        {/* PERSONAL */}
-        <nav className="space-y-1">
-          <p className="text-xs font-semibold">PERSONAL</p>
-          {EXTRA_MENUS.map(({ label, href, icon: Icon }) => (
-            <button
-              key={href}
-              onClick={() => handleNavigate(href)}
-              className="flex w-full items-center justify-between rounded-xl px-4 py-3 hover:bg-[var(--wf-highlight-soft)]"
-            >
-              <div className="flex items-center gap-4">
-                <Icon className="h-5 w-5 text-[var(--wf-subtle)]" />
-                <span>{label}</span>
-              </div>
-              <ChevronRight className="h-4 w-4 opacity-40" />
-            </button>
-          ))}
-        </nav>
-
-        <div className="my-4 border-t" />
-
-        {/* About */}
-        <nav className="space-y-1">
-          <p className="text-xs font-semibold">ABOUT</p>
-          {ABOUT_MENUS.map(({ label, href, icon: Icon }) => (
-            <button
-              key={href}
-              onClick={() => handleNavigate(href)}
-              className="flex w-full items-center justify-between rounded-xl px-4 py-3 hover:bg-[var(--wf-highlight-soft)]"
-            >
-              <div className="flex items-center gap-4">
-                <Icon className="h-5 w-5 text-[var(--wf-subtle)]" />
-                <span>{label}</span>
-              </div>
-              <ChevronRight className="h-4 w-4 opacity-40" />
-            </button>
-          ))}
-        </nav>
+ 
 
         {/* Bottom Button */}
         <div className="mt-auto pt-6">
